fix(order): the amount of a product is a decimal

+++ api/src/consumer/models.py
@@ -13,4 +13,4 @@ class Order(models.Model):
 class OrderProduct(models.Model):
     order = models.ForeignKey('Order', on_delete=models.CASCADE)
     product = models.ForeignKey('producer.Product', on_delete=models.CASCADE, related_name='+')
-    amount = models.PositiveIntegerField()
+    amount = models.DecimalField(max_digits=5, decimal_places=2)
